requirements.txt
requests>=2.31.0
anthropic>=0.34.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pandas>=2.1.0
numpy>=1.26.0
